Add files via upload
boid.js altered with the following:
Behavior0 controls the width of a bird's searching for food. When behavior0 is 0, it just looks forward. If behavior0 is 1, then the bird looks 90 degrees to the left and right
Behavior1 controls the length of a bird's searching for food. The higher the length is, the further it looks for food.
Behavior2 controls how birds react to the mouse. At 0, the birds are attracted to the mouse. At 1, the birds are repelled from the mouse.